move form config to config folder
